+++ src/GSvar/CandidateGeneDialog.cpp
@@ -74,35 +74,37 @@ void CandidateGeneDialog::updateVariants()
 		//get variants in chromosomal range
 		QList<QStringList> var_data;
 		QString af = QString::number(ui_.filter_af->value()/100.0);
-		query.exec("SELECT v.*, dvc.count_het, dvc.count_hom FROM variant v, detected_variant_counts dvc WHERE v.id=dvc.variant_id AND chr='" + chr.strNormalized(true) + "' AND start>='" + start + "' AND end<'" + end + "' AND (1000g IS NULL OR 1000g<=" + af + ") AND (exac IS NULL OR exac<=" + af + ") AND (gnomad IS NULL OR gnomad<=" + af + ") ORDER BY start");
+		QString query_text = "SELECT v.*, dvc.count_het, dvc.count_hom FROM variant v LEFT JOIN detected_variant_counts dvc ON v.id=dvc.variant_id WHERE chr='" + chr.strNormalized(true) + "' AND start>='" + start + "' AND end<'" + end + "' AND (1000g IS NULL OR 1000g<=" + af + ") AND (gnomad IS NULL OR gnomad<=" + af + ") ORDER BY start";
+		query.exec(query_text);
 		while(query.next())
 		{
 			QString var = query.value("chr").toString() + ":" + query.value("start").toString() + "-" + query.value("end").toString() + " " + query.value("ref").toString() + " > " + query.value("obs").toString();
 			QString het = query.value("count_het").toString();
 			QString hom = query.value("count_hom").toString();
 			QString gnomad = QString::number(query.value("gnomad").toDouble(), 'f', 4);
-			QString exac = QString::number(query.value("exac").toDouble(), 'f', 4);
 			QString tg = QString::number(query.value("1000g").toDouble(), 'f', 4);
-
-			//exclude somatic variants
-			if (het.toInt()+hom.toInt()==0) continue;
 
 			//filter by impact
 			QStringList parts = query.value("coding").toString().split(",");
 			QStringList parts_match;
-			foreach(QString part, parts)
+			foreach(const QString& part, parts)
 			{
 				int index = part.indexOf(':');
 				if (index==-1) continue;
 				QByteArray current_gene = part.left(index).toLatin1();
 				if (!gene_symbols.contains(current_gene)) continue;
 
-				foreach(QString impact, impacts)
+				bool match = false;
+				foreach(const QString& impact, impacts)
 				{
 					if (part.contains(impact))
 					{
-						parts_match << part;
+						match = true;
 					}
+				}
+				if (match)
+				{
+					parts_match << part;
 				}
 			}
 
@@ -112,9 +114,9 @@ void CandidateGeneDialog::updateVariants()
 				parts = parts_match;
 				parts_match.clear();
 
-				foreach(QString part, parts)
+				foreach(const QString& part, parts)
 				{
-					foreach(QString transcript_id, basic_transcripts)
+					foreach(const QString& transcript_id, basic_transcripts)
 					{
 						if (part.contains(transcript_id))
 						{
@@ -128,7 +130,7 @@ void CandidateGeneDialog::updateVariants()
 
 			//format transcript info
 			QSet<QString> types;
-			foreach(QString part, parts_match)
+			foreach(const QString& part, parts_match)
 			{
 				QStringList parts2 = part.split(":");
 				types.insert(parts2[2]);
@@ -137,7 +139,7 @@ void CandidateGeneDialog::updateVariants()
 			QString coding = parts_match.join(", ");
 
 			//add variant line to output
-			QStringList var_base = QStringList() << gene << var << het << hom << gnomad << exac << tg << type << coding;
+			QStringList var_base = QStringList() << gene << var << het << hom << gnomad << tg << type << coding;
 
 			//add sample info
 			QString var_id = query.value(0).toString();
@@ -157,7 +159,7 @@ void CandidateGeneDialog::updateVariants()
 			QMap<QString, int> hits;
 			foreach(const QStringList& line, var_data)
 			{
-				hits[line[9]] += line[11]=="hom" ? 2 : 1;
+				hits[line[9]] += line[10]=="hom" ? 2 : 1;
 			}
 
 			//remove samples with less than two hits
@@ -195,6 +197,5 @@ void CandidateGeneDialog::updateVariants()
 
 void CandidateGeneDialog::copyToClipboard()
 {
-	qDebug() << "TEST";
 	GUIHelper::copyToClipboard(ui_.variants);
 }