+++ src/GSvar/SomaticReportConfiguration.h
@@ -2,8 +2,10 @@
 #define SOMATICREPORTCONFIGURATION_H
 
 #include "ClinCnvList.h"
+#include <QBitArray>
 #include <QWidget>
 #include <QDialog>
+#include <QTableWidgetItem>
 
 namespace Ui {
 class SomaticReportConfiguration;
@@ -14,7 +16,7 @@ class SomaticReportConfiguration : public QDialog
 	Q_OBJECT
 
 public:
-	explicit SomaticReportConfiguration(const ClinCnvList& cnv_input, GeneSet keep_genes, QWidget *parent = 0);
+	explicit SomaticReportConfiguration(const ClinCnvList& cnv_input, QWidget *parent = 0);
 	~SomaticReportConfiguration();
 
 	///Returns the selected CNVs
@@ -28,8 +30,25 @@ private:
 	///gene filter for preselection of cnvs
 	GeneSet keep_genes_cnv_;
 
+	///Filter for CNVs that are shown in widget
+	QBitArray view_pass_filter;
+
 private slots:
 	void showContextMenu(QPoint pos);
+	///update widget containing CNVs after filters were adjusted
+	void filtersChanged();
+
+	///Resets view filters
+	void resetView();
+
+	///Selects variants according to view filter
+	void selectCNVsFromView();
+	///Deselects all gonosomes
+	void deselectXY();
+	///Emits signal openRegionInIGV if double click on CNV
+	void cnvDoubleClicked(QTableWidgetItem* item);
+signals:
+	void openRegionInIGV(QString region);
 };
 
 #endif // SOMATICREPORTCONFIGURATION_H