+++ src/VariantAnnotateFrequency/main.cpp
@@ -22,7 +22,7 @@ public:
 	{
 		setDescription("Annotates a variant list with variant frequencies from a BAM file.");
 		addInfile("in", "Input variant list to annotate in TSV format.", false, true);
-		addInfile("bam", "BAM file of second sample.", false, true);
+		addInfile("bam", "Input BAM file.", false, true);
 		addOutfile("out", "Output TSV file.", false, true);
 		//optional
 		addFlag("depth", "Annotate an additional column containing the depth.");